Rename `box3` to `box-3`
