+++ src/core/building.ts
@@ -5,7 +5,7 @@ import SGProp from './sgprop.js';
 import { kFileType, kFileTypeArray } from './symbols.js';
 import type Stream from './stream.js';
 import type { ConstructorOptions } from 'sc4/types';
-import Box3 from './box3.js';
+import Box3 from './box-3.js';
 import TractInfo from './tract-info.js';
 
 // # Building()
