+++ src/core/pipe.ts
@@ -10,7 +10,7 @@ import { kFileType, kFileTypeArray } from './symbols.js';
 import type { byte, dword, float, word } from 'sc4/types';
 import type Stream from './stream.js';
 import type { ConstructorOptions } from 'sc4/types';
-import Box3 from './box3.js';
+import Box3 from './box-3.js';
 import TractInfo from './tract-info.js';
 
 // # Pipe
