+++ src/core/prebuilt-network.ts
@@ -6,7 +6,7 @@ import type { byte, dword, float, word } from 'sc4/types';
 import type Stream from './stream.js';
 import { kFileType, kFileTypeArray } from './symbols.js';
 import type SGProp from './sgprop.js';
-import Box3 from './box3.js';
+import Box3 from './box-3.js';
 import TractInfo from './tract-info.js';
 
 // # PrebuiltNetwork
